requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements.txt --resolver=backtracking requirements.in
#
alembic==1.9.1
    # via dagster
aniso8601==9.0.1
    # via graphene
anyio==3.6.2
    # via
    #   starlette
    #   watchfiles
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
backoff==2.2.1
    # via gql
cachetools==5.2.1
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   dagit
    #   dagster
    #   uvicorn
coloredlogs==14.0
    # via dagster
comm==0.1.2
    # via ipykernel
contourpy==1.0.7
    # via matplotlib
croniter==1.3.8
    # via dagster
cycler==0.11.0
    # via matplotlib
dagit==1.1.7
    # via -r requirements.in
dagster==1.1.7
    # via
    #   -r requirements.in
    #   dagit
    #   dagster-graphql
    #   dagster-pandas
dagster-graphql==1.1.7
    # via dagit
dagster-pandas==0.17.7
    # via -r requirements.in
debugpy==1.6.5
    # via ipykernel
decorator==5.1.1
    # via ipython
docstring-parser==0.15
    # via dagster
entrypoints==0.4
    # via jupyter-client
executing==1.2.0
    # via stack-data
fonttools==4.38.0
    # via matplotlib
fsspec==2022.11.0
    # via universal-pathlib
google-api-core==2.11.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.16.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud==0.34.0
    # via -r requirements.in
google-cloud-core==2.3.2
    # via google-cloud-storage
google-cloud-storage==2.7.0
    # via -r requirements.in
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.0
    # via google-cloud-storage
googleapis-common-protos==1.58.0
    # via google-api-core
gql[requests]==3.4.0
    # via dagster-graphql
graphene==3.2.1
    # via dagster-graphql
graphql-core==3.2.3
    # via
    #   gql
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==2.0.1
    # via sqlalchemy
grpcio==1.47.2
    # via
    #   dagster
    #   grpcio-health-checking
grpcio-health-checking==1.43.0
    # via dagster
h11==0.14.0
    # via uvicorn
httptools==0.5.0
    # via uvicorn
humanfriendly==10.0
    # via coloredlogs
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==6.0.0
    # via alembic
importlib-resources==5.10.2
    # via alembic
ipykernel==6.20.2
    # via -r requirements.in
ipython==8.8.0
    # via ipykernel
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via dagster
joblib==1.2.0
    # via scikit-learn
jupyter-client==7.4.9
    # via ipykernel
jupyter-core==5.1.3
    # via jupyter-client
kiwisolver==1.4.4
    # via matplotlib
mako==1.2.4
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
matplotlib==3.6.3
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
multidict==6.0.4
    # via yarl
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
numpy==1.24.1
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
packaging==21.3
    # via
    #   dagster
    #   ipykernel
    #   matplotlib
pandas==1.5.2
    # via
    #   -r requirements.in
    #   dagster-pandas
    #   seaborn
parso==0.8.3
    # via jedi
pendulum==2.1.2
    # via dagster
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.4.0
    # via matplotlib
platformdirs==2.6.2
    # via jupyter-core
prompt-toolkit==3.0.36
    # via ipython
protobuf==3.20.3
    # via
    #   dagster
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-health-checking
psutil==5.9.4
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pygments==2.14.0
    # via ipython
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   croniter
    #   dagster
    #   jupyter-client
    #   matplotlib
    #   pandas
    #   pendulum
python-dotenv==0.21.0
    # via
    #   dagster
    #   uvicorn
pytz==2022.7
    # via
    #   dagster
    #   pandas
pytzdata==2020.1
    # via pendulum
pyyaml==6.0
    # via
    #   dagit
    #   dagster
    #   uvicorn
pyzmq==25.0.0
    # via
    #   ipykernel
    #   jupyter-client
requests==2.28.1
    # via
    #   dagit
    #   dagster
    #   dagster-graphql
    #   google-api-core
    #   google-cloud-storage
    #   gql
    #   requests-toolbelt
requests-toolbelt==0.10.1
    # via gql
rsa==4.9
    # via google-auth
scikit-learn==1.2.0
    # via -r requirements.in
scipy==1.10.0
    # via scikit-learn
seaborn==0.12.2
    # via -r requirements.in
setuptools==66.0.0
    # via dagster
six==1.16.0
    # via
    #   asttokens
    #   google-auth
    #   grpcio
    #   python-dateutil
sniffio==1.3.0
    # via anyio
sqlalchemy==1.4.46
    # via
    #   alembic
    #   dagster
stack-data==0.6.2
    # via ipython
starlette==0.23.1
    # via
    #   dagit
    #   dagster-graphql
tabulate==0.9.0
    # via dagster
threadpoolctl==3.1.0
    # via scikit-learn
tomli==2.0.1
    # via dagster
toposort==1.7
    # via dagster
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.64.1
    # via dagster
traitlets==5.8.1
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typing-extensions==4.4.0
    # via
    #   dagster
    #   starlette
universal-pathlib==0.0.21
    # via dagster
urllib3==1.26.13
    # via
    #   gql
    #   requests
uvicorn[standard]==0.20.0
    # via dagit
uvloop==0.17.0
    # via uvicorn
watchdog==2.2.1
    # via
    #   dagit
    #   dagster
watchfiles==0.18.1
    # via uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
websockets==10.4
    # via uvicorn
yarl==1.8.2
    # via gql
zipp==3.11.0
    # via
    #   importlib-metadata
    #   importlib-resources
